refactor: extract post form component
